requirements.txt
absl-py==1.4.0
aiofiles==23.1.0
aiohttp==3.8.5
aiohttp-cors==0.7.0
aiorwlock==1.3.0
aiosignal==1.3.1
anyio==3.7.0
appdirs==1.4.4
array-record==0.4.0
asttokens==2.2.1
astunparse==1.6.3
async-timeout==4.0.2
attrs==23.1.0
audioread==3.0.0
autoreject==0.4.2
backcall==0.2.0
beniget @ file:///dev/shm/hpcsw/SciPybundle/2022.05/foss-2022a/beniget/beniget-0.4.1
blessed==1.20.0
Bottleneck @ file:///dev/shm/hpcsw/SciPybundle/2022.05/foss-2022a/Bottleneck/Bottleneck-1.3.4
cachetools==5.3.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
cloudpickle==2.2.1
cmake==3.27.0
colorama==0.4.6
colorful==0.5.5
comm==0.1.3
contourpy==1.0.7
cycler==0.11.0
deap @ file:///dev/shm/hpcsw/SciPybundle/2022.05/foss-2022a/deap/deap-1.3.1
debugpy==1.6.7
decorator==5.1.1
distlib==0.3.7
dm-tree==0.1.8
docker-pycreds==0.4.0
dominate==2.8.0
etils==1.3.0
exceptiongroup==1.1.1
executing==1.2.0
fastapi==0.100.1
filelock==3.12.0
flatbuffers==23.5.26
fonttools==4.39.3
frozenlist==1.4.0
fsspec==2023.6.0
future==0.18.3
gast @ file:///dev/shm/hpcsw/SciPybundle/2022.05/foss-2022a/gast/gast-0.5.3
gitdb==4.0.10
GitPython==3.1.32
google-api-core==2.11.1
google-auth==2.19.1
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
googleapis-common-protos==1.59.1
gpustat==1.1
graphql-core==3.2.3
grpcio==1.54.2
h11==0.14.0
h5io==0.1.7
h5py==3.8.0
httpcore==0.17.2
httpx==0.24.1
hyperopt==0.2.7
idna==3.4
importlib-metadata==6.6.0
importlib-resources==5.12.0
ipykernel==6.23.2
ipython==8.12.2
jax==0.4.12
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.18.4
jsonschema-specifications==2023.7.1
jupyter_client==8.2.0
jupyter_core==5.3.1
keras==2.12.0
kiwisolver==1.4.4
lazy_loader==0.2
libclang==16.0.0
librosa==0.10.0.post2
lit==16.0.6
llvmlite==0.39.1
Markdown==3.4.3
MarkupSafe==2.1.2
matplotlib==3.7.1
matplotlib-inline==0.1.6
ml-dtypes==0.2.0
mne==1.3.1
mpi4py @ file:///dev/shm/hpcsw/SciPybundle/2022.05/foss-2022a/mpi4py/mpi4py-3.1.3
mpmath @ file:///dev/shm/hpcsw/SciPybundle/2022.05/foss-2022a/mpmath/mpmath-1.2.1
msgpack==1.0.5
multidict==6.0.4
nest-asyncio==1.5.6
networkx==3.1
numba==0.56.4
numexpr @ file:///dev/shm/hpcsw/SciPybundle/2022.05/foss-2022a/numexpr/numexpr-2.8.1
numpy @ file:///dev/shm/hpcsw/SciPybundle/2022.05/foss-2022a/numpy/numpy-1.22.3
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-ml-py==12.535.77
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
oauthlib==3.2.2
opencensus==0.11.2
opencensus-context==0.1.3
opencv-python==4.8.0.74
opt-einsum==3.3.0
packaging==23.1
pandas @ file:///dev/shm/hpcsw/SciPybundle/2022.05/foss-2022a/pandas/pandas-1.4.2
parso==0.8.3
pathtools==0.1.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==3.2.0
ply @ file:///dev/shm/hpcsw/SciPybundle/2022.05/foss-2022a/ply/ply-3.11
pooch==1.6.0
prometheus-client==0.17.1
promise==2.3
prompt-toolkit==3.0.38
protobuf==4.23.2
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
py-spy==0.3.14
py4j==0.10.9.7
pyarrow==12.0.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
pybind11 @ file:///dev/shm/hpcsw/pybind11/2.9.2/GCCcore-11.3.0/pybind11-2.9.2
pycparser==2.21
pydantic==1.10.12
pydub==0.25.1
Pygments==2.15.1
pymatreader==0.0.31
pyparsing==3.0.9
python-dateutil==2.8.2
pythran @ file:///dev/shm/hpcsw/SciPybundle/2022.05/foss-2022a/pythran/pythran-0.11.0
pytorch-ignite==0.4.12
pytz==2023.3
PyYAML==6.0
pyzmq==25.1.0
ray==2.6.1
referencing==0.30.0
requests==2.28.2
requests-oauthlib==1.3.1
rpds-py==0.9.2
rsa==4.9
scikit-learn==1.2.2
SciPy @ file:///dev/shm/hpcsw/SciPybundle/2022.05/foss-2022a/scipy/scipy-1.8.1
sentry-sdk==1.28.1
setproctitle==1.3.2
sgqlc==16.2
six==1.16.0
smart-open==6.3.0
smmap==5.0.0
sniffio==1.3.0
soundfile==0.12.1
soxr==0.3.5
stack-data==0.6.2
starlette==0.27.0
sympy==1.11.1
tensorboard==2.12.3
tensorboard-data-server==0.7.0
tensorboardX==2.6.2
termcolor==2.3.0
threadpoolctl==3.1.0
toml==0.10.2
torch==1.13.1+cu116
torchaudio==0.13.1+cu116
torchvision==0.14.1+cu116
tornado==6.3.2
tqdm==4.65.0
traitlets==5.9.0
triton==2.0.0
typing_extensions==4.5.0
tzdata==2023.3
urllib3==1.26.15
uvicorn==0.23.2
virtualenv==20.21.0
visdom==0.2.4
wandb==0.15.5
wcwidth==0.2.6
websocket-client==1.6.1
Werkzeug==2.3.6
wrapt==1.14.1
xmltodict==0.13.0
yarl==1.9.2
zipp==3.15.0
